requirements.txt
access==1.1.3
affine==2.3.0
amply==0.1.4
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
coverage==5.5
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-decouple==2.1
django-on-heroku==1.1.2
django-taggit==1.3.0
docutils==0.16
esda==2.3.6
Fiona==1.8.18
folium==0.12.1
geographiclib==1.50
geopandas==0.9.0
geopy==2.1.0
giddy==2.3.3
gunicorn==20.1.0
idna==2.10
inequality==1.0.0
iniconfig==1.1.1
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
libpysal==4.4.0
llvmlite==0.36.0
mapclassify==2.4.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mgwr==2.1.2
mpmath==1.2.1
munch==2.5.0
networkx==2.5
numba==0.53.0
numpy==1.20.1
opencv-contrib-python==4.5.1.48
packaging==20.9
pandas==1.2.3
patsy==0.5.1
Pillow==8.1.2
pluggy==0.13.1
pointpats==2.2.0
psycopg2-binary==2.8.6
PuLP==2.4
py==1.10.0
pygeos==0.9
pyparsing==2.4.7
pyproj==3.0.1
pysal==2.4.0
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.0
python-dotenv==0.17.0
pytz==2021.1
quantecon==0.4.8
rasterio==1.2.1
rasterstats==0.15.0
requests==2.25.1
Rtree==0.9.7
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
segregation==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
snuggs==1.4.7
soupsieve==2.2
spaghetti==1.5.6
spglm==1.0.8
spint==1.0.7
splot==1.1.3
spopt==0.1.1
spreg==1.2.2
spvcm==0.3.0
sqlparse==0.4.1
static3==0.7.0
statsmodels==0.12.2
sympy==1.7.1
threadpoolctl==2.1.0
tobler==0.7.0
toml==0.10.2
tqdm==4.59.0
urllib3==1.26.4
whitenoise==5.2.0
